Typecheck, Lint and Test with Github Action
I updated the README.md to replace the old repo name with
blue-core-lod/bluecore-workflows. I also updated it to reflect that we
have replaced poetry with uv. I think there are other things in their
that might need updating as things settle.

This commit includes the result of running ruff format. There is still
an open question about whether we want to do flake8 testing in CI.

Closes #2