Airflow 2-to-3
This PR includes lots of changes to upgrade Airflow from v2 to v3. The
primary changes were to the docker-compose.yaml file which is now named
compose.yaml now includes new containers for:

- airflow-apiserver
- airflow-dag-processor
- airflow-triggerer

Since we are using a different compose setup in blue-core-lod/terraform
there is now just one Docker compose configuration file here and it is
meant for use in development of bluecore-workflows and also bluecore_api.

Some Operators moved around into `airflow.providers.standard.operators`.
Some functions moved into `airflow.sdk`.

Note that xmlsec needed to be pinned to v1.3.14 because there's an issue
with v1.3.15 when installing with uv:

```
   [stderr]
  WARNING setuptools_scm.pyproject_reading toml section missing
  'pyproject.toml does not contain a tool.setuptools_scm section'
  Traceback (most recent call last):
    File
  "/home/runner/.cache/uv/builds-v0/.tmpMMJuUN/lib/python3.12/site-packages/setuptools_scm/_integration/pyproject_reading.py",
  line 36, in read_pyproject
      section = defn.get("tool", {})[tool_name]
                ~~~~~~~~~~~~~~~~~~~~^^^^^^^^^^^
  KeyError: 'setuptools_scm'
  /home/runner/.cache/uv/builds-v0/.tmpMMJuUN/lib/python3.12/site-packages/setuptools/dist.py:759:
  SetuptoolsDeprecationWarning: License classifiers are deprecated.
```