Use save_graph for persisting to the database
Use the new `save_graph()` function in bluecore_models for persisting a
rdflib Graph to the databae. This will handle updating resources as
needed. Testing of this functionality lives over in the bluecore-models repository.

Note: setting `AIRFLOW_VAR_BLUECORE_URL` is needed for setting the default
Bluecore namespace. For development, the namespace http://localhost:3000
is used since it is assumed you are running bluecore_api on port 3000
in development. In production we will be using `http://dev.bcld.info".

Closes #67